requirements.txt
opencv-python
numpy
faster-whisper
ffmpeg-python
rich
requests
python-dotenv
